pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "garmin-cycling-coach"
version = "0.1.0"
description = "Personal Garmin cycling coach agent"
requires-python = ">=3.11"
dependencies = [
    "garminconnect>=0.2.25",
    "python-dotenv>=1.0.1",
    "openai>=1.40.0",
    "rich>=13.7.0",
]

[project.scripts]
garmin-coach = "garmin_coach.cli:main"
garmin-coach-web = "garmin_coach.web:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
